split client and server modules
