+++ lib/level10/games.ex
@@ -1,39 +1,60 @@
 defmodule Level10.Games do
   @moduledoc """
-  This context module handles all of the work around running games. Most of the
-  functions will take in a game struct and manipulate that struct and return
-  it.
+  This module is the interface into game logic. All presenters within the web
+  domain should interface only with this module for controlling games. Its
+  children shouldn't be touched directly.
+
+  Most of the functions in the module are client functions that give
+  instructions to a game server, but some of them will interact instead with
+  the distributed registry and supervisor, or with Phoenix Presence or PubSub.
   """
 
-  alias Level10.Games.{Card, Game, GameServer, Levels, Player}
+  alias Level10.Presence
+  alias Level10.Games.{Game, GameRegistry, GameServer, GameSupervisor, Levels, Player}
   require Logger
 
-  @typep event_type :: atom()
+  @typep game_name :: {:via, module, term}
+
+  @max_creation_attempts 10
 
   @doc """
   Add one or more cards to a group that is already on the table
   """
   @spec add_to_table(Game.join_code(), Player.id(), Player.id(), non_neg_integer(), Game.cards()) ::
           :ok | :invalid_group | :level_incomplete | :needs_to_draw | :not_your_turn
-  defdelegate add_to_table(join_code, player_id, table_id, position, cards_to_add), to: GameServer
+  def add_to_table(join_code, player_id, table_id, position, cards_to_add) do
+    GenServer.call(
+      via(join_code),
+      {:add_to_table, {player_id, table_id, position, cards_to_add}},
+      5000
+    )
+  end
 
   @doc """
   Get the current count of active games in play.
   """
   @spec count() :: non_neg_integer()
-  defdelegate count, to: GameServer
+  def count do
+    %{active: count} = Supervisor.count_children(GameSupervisor)
+    count
+  end
 
   @doc """
   Create a new game with the player named as its creator.
   """
   @spec create_game(String.t()) :: {:ok, Game.join_code(), Player.id()} | :error
-  defdelegate create_game(player_name), to: GameServer
+  def create_game(player_name) do
+    player = Player.new(player_name)
+    do_create_game(player, @max_creation_attempts)
+  end
 
   @doc """
   Returns a Player struct representing the player who created the game.
   """
   @spec creator(Game.join_code()) :: Player.t()
-  defdelegate creator(join_code), to: GameServer
+  def creator(join_code) do
+    GenServer.call(via(join_code), :creator, 5000)
+  end
 
   @doc """
   Check to see if the current player has drawn a card yet.
@@ -44,7 +65,9 @@ defmodule Level10.Games do
       true
   """
   @spec current_player_has_drawn?(Game.join_code()) :: boolean()
-  defdelegate current_player_has_drawn?(join_code), to: GameServer
+  def current_player_has_drawn?(join_code) do
+    GenServer.call(via(join_code), :current_turn_drawn?, 5000)
+  end
 
   @doc """
   Delete a game.
@@ -54,8 +77,10 @@ defmodule Level10.Games do
       iex> delete_game("ABCD")
       :ok
   """
-  @spec delete_game(Game.join_code()) :: :ok
-  defdelegate delete_game(join_code), to: GameServer
+  @spec delete_game(Game.join_code(), reason :: term, timeout) :: :ok
+  def delete_game(join_code, reason \\ :normal, timeout \\ :infinity) do
+    GenServer.stop(via(join_code), reason, timeout)
+  end
 
   @doc """
   Discard a card from the player's hand
@@ -67,7 +92,9 @@ defmodule Level10.Games do
   """
   @spec discard_card(Game.join_code(), Player.id(), Card.t()) ::
           :ok | :needs_to_draw | :not_your_turn
-  defdelegate discard_card(join_code, player_id, card), to: GameServer
+  def discard_card(join_code, player_id, card) do
+    GenServer.call(via(join_code), {:discard, {player_id, card}}, 5000)
+  end
 
   @doc """
   Take the top card from either the draw pile or discard pile and add it to the
@@ -83,7 +110,9 @@ defmodule Level10.Games do
   """
   @spec draw_card(Game.join_code(), Player.id(), :discard_pile | :draw_pile) ::
           Card.t() | :already_drawn | :empty_discard_pile | :not_your_turn | :skip
-  defdelegate draw_card(join_code, player_id, source), to: GameServer
+  def draw_card(join_code, player_id, source) do
+    GenServer.call(via(join_code), {:draw, {player_id, source}}, 5000)
+  end
 
   @doc """
   Returns whether or not a game with the specified join code exists.
@@ -97,10 +126,25 @@ defmodule Level10.Games do
       false
   """
   @spec exists?(Game.join_code()) :: boolean()
-  defdelegate exists?(join_code), to: GameServer
+  def exists?(join_code) do
+    case Horde.Registry.lookup(GameRegistry, join_code) do
+      [] -> false
+      _ -> true
+    end
+  end
 
+  @doc """
+  Returns whether or not the specified game is finished.
+
+  ## Examples
+
+      iex> finished?("ABCD")
+      true
+  """
   @spec finished?(Game.join_code()) :: boolean()
-  defdelegate finished?(join_code), to: GameServer
+  def finished?(join_code) do
+    GenServer.call(via(join_code), :finished?, 5000)
+  end
 
   @doc """
   Returns the game with the specified join code.
@@ -111,7 +155,9 @@ defmodule Level10.Games do
       %Game{}
   """
   @spec get(Game.join_code()) :: Game.t()
-  defdelegate get(join_code), to: GameServer
+  def get(join_code) do
+    GenServer.call(via(join_code), :get, 5000)
+  end
 
   @doc """
   Get the player whose turn it currently is.
@@ -122,7 +168,9 @@ defmodule Level10.Games do
       %Player{id: "ffe6629a-faff-4053-b7b8-83c3a307400f", name: "Player 1"}
   """
   @spec get_current_turn(Game.join_code()) :: Player.t()
-  defdelegate get_current_turn(join_code), to: GameServer
+  def get_current_turn(join_code) do
+    GenServer.call(via(join_code), :current_player, 5000)
+  end
 
   @doc """
   Get the count of cards in each player's hand.
@@ -133,7 +181,9 @@ defmodule Level10.Games do
       %{"179539f0-661e-4b56-ac67-fec916214223" => 10, "000cc69a-bb7d-4d3e-ae9f-e42e3dcac23e" => 3}
   """
   @spec get_hand_counts(Game.join_code()) :: %{optional(Player.id()) => non_neg_integer()}
-  defdelegate get_hand_counts(join_code), to: GameServer
+  def get_hand_counts(join_code) do
+    GenServer.call(via(join_code), :hand_counts, 5000)
+  end
 
   @doc """
   Get the hand of the specified player.
@@ -144,7 +194,9 @@ defmodule Level10.Games do
       [%Card{color: :green, value: :twelve}, %Card{color: :blue, value: :nine}, ...]
   """
   @spec get_hand_for_player(Game.join_code(), Player.id()) :: list(Card.t())
-  defdelegate get_hand_for_player(join_code, player_id), to: GameServer
+  def get_hand_for_player(join_code, player_id) do
+    GenServer.call(via(join_code), {:hand, player_id}, 5000)
+  end
 
   @doc """
   Get the level information for each player in the game.
@@ -158,7 +210,13 @@ defmodule Level10.Games do
       }
   """
   @spec get_levels(Game.join_code()) :: %{optional(Player.t()) => Levels.level()}
-  defdelegate get_levels(join_code), to: GameServer
+  def get_levels(join_code) do
+    levels = GenServer.call(via(join_code), :levels, 5000)
+
+    for {player_id, level_number} <- levels,
+        into: %{},
+        do: {player_id, Levels.by_number(level_number)}
+  end
 
   @doc """
   Get the list of players in a game.
@@ -172,16 +230,25 @@ defmodule Level10.Games do
       ]
   """
   @spec get_players(Game.join_code()) :: list(Player.t())
-  defdelegate get_players(join_code), to: GameServer
+  def get_players(join_code) do
+    GenServer.call(via(join_code), :players, 5000)
+  end
 
   @doc """
   Gets the set of IDs of players who are ready for the next round to begin.
   """
   @spec get_players_ready(Game.join_code()) :: MapSet.t(Player.id())
-  defdelegate get_players_ready(join_code), to: GameServer
+  def get_players_ready(join_code) do
+    GenServer.call(via(join_code), :players_ready, 5000)
+  end
 
+  @doc """
+  Get the round number for the current round.
+  """
   @spec get_round_number(Game.join_code()) :: non_neg_integer()
-  defdelegate get_round_number(join_code), to: GameServer
+  def get_round_number(join_code) do
+    GenServer.call(via(join_code), :current_round, 5000)
+  end
 
   @doc """
   Get the scores for all players in a game.
@@ -195,7 +262,9 @@ defmodule Level10.Games do
       }
   """
   @spec get_scores(Game.join_code()) :: Game.scores()
-  defdelegate get_scores(join_code), to: GameServer
+  def get_scores(join_code) do
+    GenServer.call(via(join_code), :scoring, 5000)
+  end
 
   @doc """
   Get the table: the cards that have been played to complete levels by each
@@ -220,7 +289,9 @@ defmodule Level10.Games do
       }
   """
   @spec get_table(Game.join_code()) :: Game.table()
-  defdelegate get_table(join_code), to: GameServer
+  def get_table(join_code) do
+    GenServer.call(via(join_code), :table, 5000)
+  end
 
   @doc """
   Get the top card from the discard pile.
@@ -234,7 +305,9 @@ defmodule Level10.Games do
       nil
   """
   @spec get_top_discarded_card(Game.join_code()) :: Card.t() | nil
-  defdelegate get_top_discarded_card(join_code), to: GameServer
+  def get_top_discarded_card(join_code) do
+    GenServer.call(via(join_code), :top_discarded_card, 5000)
+  end
 
   @doc """
   Attempts to join a game. Will return an ok tuple with the player ID for the
@@ -256,7 +329,15 @@ defmodule Level10.Games do
   """
   @spec join_game(Game.join_code(), String.t()) ::
           {:ok, Player.id()} | :already_started | :full | :not_found
-  defdelegate join_game(join_code, player_name), to: GameServer
+  def join_game(join_code, player_name) do
+    player = Player.new(player_name)
+
+    if exists?(join_code) do
+      GenServer.call(via(join_code), {:join, player}, 5000)
+    else
+      :not_found
+    end
+  end
 
   @doc """
   Removes the specified player from the game. This is only allowed if the game
@@ -266,20 +347,38 @@ defmodule Level10.Games do
   well.
   """
   @spec leave_game(Game.join_code(), Player.id()) :: :ok | :already_started | :deleted
-  defdelegate leave_game(join_code, player_id), to: GameServer
+  def leave_game(join_code, player_id) do
+    result = GenServer.call(via(join_code), {:delete_player, player_id}, 5000)
+    with :empty_game <- result, do: delete_game(join_code)
+  end
+
+  @doc """
+  Get the list of players currently present in the specified game.
+  """
+  @spec list_presence(Game.join_code()) :: %{optional(Player.id()) => map()}
+  def list_presence(join_code) do
+    Presence.list("game:" <> join_code)
+  end
 
   @doc """
   Stores in the game state that the specified player is ready to move on to the
   next stage of the game.
   """
   @spec mark_player_ready(Game.join_code(), Player.id()) :: :ok
-  defdelegate mark_player_ready(join_code, player_id), to: GameServer
+  def mark_player_ready(join_code, player_id) do
+    result = GenServer.call(via(join_code), {:player_ready, player_id}, 5000)
+    with :game_over <- result, do: delete_game(join_code)
+  end
 
   @doc """
   Returns whether or not the specified player exists within the specified game.
   """
   @spec player_exists?(Game.t() | Game.join_code(), Player.id()) :: boolean()
-  defdelegate player_exists?(join_code, player_id), to: GameServer
+  def player_exists?(join_code, player_id) when is_binary(join_code) do
+    GenServer.call(via(join_code), {:player_exists?, player_id}, 5000)
+  end
+
+  def player_exists?(game, player_id), do: Game.player_exists?(game, player_id)
 
   @doc """
   Check whether or not the current round has started.
@@ -293,25 +392,33 @@ defmodule Level10.Games do
       false
   """
   @spec round_started?(Game.join_code()) :: boolean()
-  defdelegate round_started?(join_code), to: GameServer
+  def round_started?(join_code) do
+    GenServer.call(via(join_code), :round_started?, 5000)
+  end
 
   @doc """
   Returns the player struct representing the player who won the current round.
   """
   @spec round_winner(Game.join_code()) :: Player.t() | nil
-  defdelegate round_winner(join_code), to: GameServer
+  def round_winner(join_code) do
+    GenServer.call(via(join_code), :round_winner, 5000)
+  end
 
   @doc """
   Start the next round.
   """
   @spec start_round(Game.join_code()) :: :ok | :game_over
-  defdelegate start_round(join_code), to: GameServer
+  def start_round(join_code) do
+    GenServer.call(via(join_code), :start_round, 5000)
+  end
 
   @doc """
   Start the game.
   """
   @spec start_game(Game.join_code()) :: :ok | :single_player
-  defdelegate start_game(join_code), to: GameServer
+  def start_game(join_code) do
+    GenServer.call(via(join_code), :start_game, 5000)
+  end
 
   @doc """
   Check whether or not a game has started.
@@ -325,42 +432,84 @@ defmodule Level10.Games do
       false
   """
   @spec started?(Game.join_code()) :: boolean()
-  defdelegate started?(join_code), to: GameServer
+  def started?(join_code) do
+    GenServer.call(via(join_code), :started?, 5000)
+  end
+
+  @doc """
+  Susbscribe a process to updates for the specified game.
+  """
+  @spec subscribe(String.t(), Player.id()) :: :ok | {:error, term()}
+  def subscribe(game_code, player_id) do
+    topic = "game:" <> game_code
+
+    with :ok <- Phoenix.PubSub.subscribe(Level10.PubSub, topic),
+         {:ok, _} <- Presence.track_player(game_code, player_id) do
+      Presence.track_user(player_id, game_code)
+      :ok
+    end
+  end
 
   @doc """
   Set the given player's table to the given cards.
   """
   @spec table_cards(Game.join_code(), Player.id(), Game.player_table()) ::
           :ok | :already_set | :needs_to_draw | :not_your_turn
-  defdelegate table_cards(join_code, player_id, player_table), to: GameServer
-
-  @doc """
-  Susbscribe a process to updates for the specified game.
-  """
-  @spec subscribe(String.t(), Player.id()) :: :ok | {:error, term()}
-  defdelegate subscribe(game_code, player_id), to: GameServer
+  def table_cards(join_code, player_id, player_table) do
+    GenServer.call(via(join_code), {:table_cards, {player_id, player_table}}, 5000)
+  end
 
   @doc """
   Unsubscribe a process from updates for the specified game.
   """
   @spec unsubscribe(String.t(), Player.id()) :: :ok | {:error, term()}
-  defdelegate unsubscribe(game_code, player_id), to: GameServer
+  def unsubscribe(game_code, player_id) do
+    topic = "game:" <> game_code
+
+    with :ok <- Phoenix.PubSub.unsubscribe(Level10.PubSub, topic) do
+      Presence.untrack(self(), topic, player_id)
+    end
+  end
 
   @doc """
   Update the specified game using the provided function.
   """
   @spec update(Game.join_code(), (Game.t() -> Game.t())) :: :ok
-  defdelegate update(join_code, fun), to: GameServer
+  def update(join_code, fun) do
+    GenServer.call(via(join_code), {:update, fun}, 5000)
+  end
 
-  @doc """
-  Send an update to all the subscribed processes
-  """
-  @spec broadcast(Game.join_code(), event_type(), term()) :: :ok | {:error, term()}
-  defdelegate broadcast(join_code, event_type, event), to: GameServer
+  # Private
 
-  @doc """
-  Get the list of players currently present in the specified game.
-  """
-  @spec list_presence(Game.join_code()) :: %{optional(Player.id()) => map()}
-  defdelegate list_presence(join_code), to: GameServer
+  @spec do_create_game(Player.t(), non_neg_integer()) ::
+          {:ok, Game.join_code(), Player.id()} | :error
+  defp do_create_game(player, attempts_remaining)
+
+  defp do_create_game(_player, 0) do
+    :error
+  end
+
+  defp do_create_game(player, attempts_remaining) do
+    join_code = Game.generate_join_code()
+
+    game = %{
+      id: join_code,
+      start: {GameServer, :start_link, [{join_code, player}, [name: via(join_code)]]},
+      restart: :temporary
+    }
+
+    case Horde.DynamicSupervisor.start_child(GameSupervisor, game) do
+      {:ok, _pid} ->
+        Logger.info(["Created game ", join_code])
+        {:ok, join_code, player.id}
+
+      {:error, {:already_started, _pid}} ->
+        do_create_game(player, attempts_remaining - 1)
+    end
+  end
+
+  @spec via(Game.join_code()) :: game_name()
+  defp via(join_code) do
+    {:via, Horde.Registry, {GameRegistry, join_code}}
+  end
 end
